requirements.txt
azure==0.11.1
Django==1.8.3
django-azure-storage==0.0.3
futures==3.0.3
numpy==1.9.2
opencv-python==2.4.11
Pillow==2.9.0
python-dateutil==2.4.2
six==1.9.0
